Completed the lab
Completed the lab. The implementation of AVL Tree didn't tree didn't
took much long but a typo caused me a lot of time to fix and make the
code work correctly. Also much of the code is alike GeeksForGeeks since
I completed the lab with the help of their article on AVL Trees